Create futures_historical_data_retrieval.ipynb
Binance Futures API to get historical klines (candlestick) data for a given symbol and interval. The API endpoint returns the data in a JSON format, which we then parse into a pandas DataFrame. We set the index of the DataFrame to be the "Open time" column, which is converted to a datetime format.